fix(tts): Refactor report generating into class
